backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-multipart
pydantic
joblib
scikit-learn
pandas
numpy
passlib[bcrypt]
python-jose[cryptography]
requests
torch
torchvision
Pillow
ultralytics
